Adicionando as questões 13 ate 15 e modificando a 11 e 12

+++ capitulo-5/exercicio5.11.py
@@ -1,5 +1,6 @@
-deposito = float(input('Digite o deposito inicial: '))
-taxa = float(input('Digite a taxa de juros: '))
+from decimal import Decimal
+deposito = Decimal(input('Digite o deposito inicial: '))
+taxa = Decimal(input('Digite a taxa de juros: '))
 acumulador = deposito
 mes = 1
 while(mes <= 24):
